+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Adjust Counter Barrage rocket spawn workaround
 - Minor Aspect of Hel now blocks all omega Attacks and Specials during Valkyrie form and properly resumes charging them after it ends.
 - Remove ImGui menu, its been outdated and unnecessary since the in-game UI was introduced.
+- Fix Pan plus Morrigan not having homing Specials
 
 ## [0.9.6] - 2026-08-04